refactor transactions view and styling
